Modelo y rutas de Perido terminadas
Pendiente vinculos con altas_sga

+++ database/models/altasSGA.model.js
@@ -30,7 +30,7 @@ const AltaSGASchema = {
     },
     unidades:{
         allowNull: false,
-        type: DataTypes.DOUBLE(5, 2),
+        type: DataTypes.FLOAT,
         field: 'unidades',
     },
     oficio:{
